Turn `Facets` into a separate component
