+++ src/components/search/Filter/Filter.tsx
@@ -4,88 +4,14 @@ import type {
   IStoreSelectedFacet,
   FacetedFilter_FacetsFragment,
 } from '@generated/graphql'
-import { List as UIList, Label as UILabel } from '@faststore/ui'
 import useWindowDimensions from 'src/hooks/useWindowDimensions'
 import Button from 'src/components/ui/Button'
-import Checkbox from 'src/components/ui/Checkbox'
-import { Badge } from 'src/components/ui/Badge'
 import { X as XIcon } from 'phosphor-react'
-import Accordion, { AccordionItem } from 'src/components/ui/Accordion'
 import SlideOver from 'src/components/ui/SlideOver'
 
+import Facets from './Facets'
+
 import './filter.scss'
-
-interface FacetsProps {
-  slug: string
-  testId: string
-  selectedFacets: IStoreSelectedFacet[]
-  filteredFacets: FacetedFilter_FacetsFragment[]
-  indicesExpanded: Set<number>
-  onSelectFacet: ({ key, value }: IStoreSelectedFacet) => void
-  onAccordionChange: (index: number) => void
-  onAccordionItemMount: (
-    index: number,
-    values: FacetedFilter_FacetsFragment['values']
-  ) => void
-}
-
-const Facets = ({
-  slug,
-  testId,
-  selectedFacets,
-  filteredFacets,
-  indicesExpanded,
-  onSelectFacet,
-  onAccordionChange,
-  onAccordionItemMount,
-}: FacetsProps) => {
-  return (
-    <div className="filter" data-store-filter data-testid={testId}>
-      <Accordion expandedIndices={indicesExpanded} onChange={onAccordionChange}>
-        {filteredFacets.map(({ label, values, key }, index) => (
-          <AccordionItem
-            key={`${label}-${index}`}
-            testId="filter-accordion"
-            isExpanded={indicesExpanded.has(index)}
-            buttonLabel={label}
-            ref={(_) => onAccordionItemMount(index, values)}
-          >
-            <UIList>
-              {values.map((item) => {
-                const id = `${label}-${item.label}`
-
-                return (
-                  <li key={id} className="filter__item">
-                    <Checkbox
-                      id={id}
-                      checked={
-                        item.value === slug ||
-                        selectedFacets.some(
-                          (facet) => facet.value === item.value
-                        )
-                      }
-                      onChange={() => onSelectFacet({ key, value: item.value })}
-                      data-testid="filter-accordion-panel-checkbox"
-                      data-value={item.value}
-                      data-quantity={item.quantity}
-                      disabled={item.value === slug}
-                    />
-                    <UILabel htmlFor={id} className="title-small">
-                      {item.label}{' '}
-                      <Badge variant="neutral" small>
-                        {item.quantity}
-                      </Badge>
-                    </UILabel>
-                  </li>
-                )
-              })}
-            </UIList>
-          </AccordionItem>
-        ))}
-      </Accordion>
-    </div>
-  )
-}
 
 interface FilterProps {
   facets: FacetedFilter_FacetsFragment[]
